huge step forward
- Implemented evaluation of fitted data
- Re-implemented compile-time dynamic partitions with tf.map_fn
- Removed dependency on design_matrix inside tensorflow graph;
    however it is still included
- Added compile-time specification of shapes
TODO: Wait for tf.unique_v2 to be included in official release

+++ impl/tf/linear_regression/linear_regression.py
@@ -25,7 +25,7 @@ class LinearRegression:
             :return:    | tuple(b, least_squares)
                         | b is a Tensor of shape ([...], N, K)
             """
-        l2_reg = tf.convert_to_tensor(l2_reg, name="l2_reg")
+        l2_reg = tf.convert_to_tensor(l2_reg, dtype=X.dtype, name="l2_reg")
         # lambda_I = tf.tile(l2_reg, (tf.shape(X)[-2], tf.shape(X)[-2]))
         
         with tf.name_scope(name):
